pyproject.toml
# --- PROJECT CONFIGURATION --- #

# Metadata (see https://peps.python.org/pep-0621/)
[project]
name = "nged-substation-forecast"
version = "0.0.1"
description = "Forecast net demand at primary, BSP, and GSP substations, and split the forecast into gross demand, solar, and wind per substation for National Grid Electricity Distribution (NGED)"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.14"
license = {text = "MIT License"}
authors = [
    { name = "Open Climate Fix Team", email = "info@openclimatefix.org"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "dagster>=1.12.11",
    "dagster-webserver>=1.12.11",
    "patito",
    "polars>=1.0.0",
    "requests>=2.31.0",
    "mlflow-skinny",
    "dynamical_data",
    "nged-data",
    "contracts",
    "ml_core",
    "obstore",
    "xgboost_forecaster",
    "hydra-core",
    "deltalake",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "dagster-dg-cli",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.13",
    "ty>=0.0.16",
    "unittest-xml-reporting>=4.0.0",
    "hypothesis>=6.0.0",
]

[project.urls]
repository = "https://github.com/openclimatefix/nged-substation-forecast"

# --- UV WORKSPACE CONFIG ---
# See: https://docs.astral.sh/uv/concepts/projects/workspaces

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
nged-data = { workspace = true }
contracts = { workspace = true }
ml_core = { workspace = true }
xgboost_forecaster = { workspace = true }
dynamical_data = { workspace = true }

# --- DAGSTER ---

[tool.dagster]
module_name = "nged_substation_forecast.definitions"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "nged_substation_forecast"
registry_modules = [
    "nged_substation_forecast.components.*",
]

# --- LINTING AND TYPING CONFIGURATION --- #

# Ruff configuration
# See https://docs.astral.sh/ruff/configuration/
[tool.ruff]
line-length = 100

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = 100
